Inventory file
in our class we can creat a sub class for variables. So in our class called Public say we want some bullets. Our sub class could then be called Stuff, and in this sub class we can also decide how many bullets we can have. 

() = a Contructor, which basically allows you to giv\e defult values in your class and actions. Another thing is the constructors name is always going to be the same as the name of the class (ex. Stuff) 
Constructos will always match up with a Parameter